+++ src/math_custom.c
@@ -1,11 +1,10 @@
 #include "./math_custom.h"
 
 #include <math.h>
-#include <stdio.h>
 #include <stdlib.h>
 
 inline float transform3Dto2D(float xy, float z) {
-  return xy / ((z)*tan(HALF_FOV_ANGLE_RADIANS));
+  return xy / ((z)*(float)tan(HALF_FOV_ANGLE_RADIANS));
 }
 
 inline float randF(float min, float max) {
